+++ test/src/Bucket/BucketBasicOperationTest.cc
@@ -346,7 +346,7 @@ TEST_F(BucketBasicOperationTest, ListBucketspagingTest)
     //list by step
     request.setMaxKeys(2);
     bool IsTruncated = false;
-    int total = 0;
+    size_t total = 0;
     do {
         outcome = Client->ListBuckets(request);
         EXPECT_EQ(outcome.isSuccess(), true);
@@ -355,7 +355,7 @@ TEST_F(BucketBasicOperationTest, ListBucketspagingTest)
         request.setMarker(outcome.result().NextMarker());
         IsTruncated = outcome.result().IsTruncated();
     } while (IsTruncated);
-    EXPECT_EQ(total, 5);
+    EXPECT_EQ(total, 5UL);
 
     //delete all
     request.setMaxKeys(100);
